+++ api/routes/auth/roles/get.endpoint.js
@@ -1,0 +1,30 @@
+const {
+  getDB,
+  getFullPath,
+  login,
+  request,
+  unauthenticatedTest,
+  unauthorizedTest
+} = require('../../../utils.endpoint');
+
+describe('auth roles endpoint | GET /auth/roles', async () => {
+  const db = getDB();
+  beforeAll(() => db.seed.run());
+  afterAll(() => db.destroy());
+
+  const url = getFullPath('/auth/roles');
+
+  unauthenticatedTest('get', url);
+  unauthorizedTest('get', url);
+
+  it('when authenticated', async () => {
+    const cookies = await login();
+    const { response, body } = await request.get(url, {
+      jar: cookies,
+      json: true
+    });
+
+    expect(response.statusCode).toEqual(200);
+    expect(body).toMatchSnapshot();
+  });
+});
